% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aVirtualTwins.R
\docType{package}
\name{aVirtualTwins}
\alias{aVirtualTwins}
\alias{aVirtualTwins-package}
\title{aVirtualTwins : An adapation of VirtualTwins method created by Jared Foster.}
\description{
aVirtualTwins is written mainly with reference classes. Briefly, there is three kinds of class :
\itemize{
 \item \code{\link{VT.object}} class to represent RCT dataset used by aVirtualTwins. To format correctly RCT dataset, use \code{\link{formatRCTDataset}}.
 \item \code{\link{VT.difft}} class to compute difference between twins. Family \code{\link{VT.forest}} extends it to compute twins by random forest.
 \item \code{\link{VT.tree}} class to find subgroups from \code{difft} by CART trees. \code{\link{VT.tree.class}} and \code{\link{VT.tree.reg}} extend it.
}
}
\section{TODO LIST}{

\emph{last update : 24.07.2015}
\itemize{
 \item More detailed documentation and vignettes
 \item Write wrappers for classes
 \item Write examples
 \item ...
}
}

