\name{VirtualTwins-package}
\alias{VirtualTwins-package}
\alias{VirtualTwins}
\docType{package}
\title{
\packageTitle{VirtualTwins}
}
\description{
\packageDescription{VirtualTwins}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VirtualTwins}
\packageIndices{VirtualTwins}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VirtualTwins}

Maintainer: \packageMaintainer{VirtualTwins}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
