% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/difft.R
\docType{class}
\name{VT.difft-class}
\alias{VT.difft}
\alias{VT.difft-class}
\title{A reference class to represent difference between twin1 and twin2}
\description{
A reference class to represent difference between twin1 and twin2
}
\section{Fields}{

\describe{
\item{\code{vt.object}}{VT.object (refClass) representing data}

\item{\code{twin1}}{vector of \eqn{E(Y|T= real treatment)}}

\item{\code{twin2}}{vector of \eqn{E(Y|T= antoher treatment)}}

\item{\code{difft}}{vector of difference between twin1 and twin2}
}}
\section{Methods}{

\describe{
\item{\code{computeDifft()}}{Compute difference between twin1 and twin2}
}}

