% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/difft.R
\docType{class}
\name{VT.difft}
\alias{VT.difft}
\title{Difference between twins}
\description{
A reference class to represent difference between twin1 and twin2
}
\details{
Difft are calculated depending on the favorable outcome chosen. It is the
second level of the outcome. For example, if the outcome is 0 and 1, the
favorable outcome is 1. Then, \deqn{difft_i = twin1_i - twin2_i IF T_i =
1} \deqn{ difft_i = twin2_i - twin1_i IF T_i = 0}
}
\section{Fields}{

\describe{
\item{\code{vt.object}}{VT.object (refClass) representing data}

\item{\code{twin1}}{vector of \eqn{E(Y|T = real treatment)}}

\item{\code{twin2}}{vector of \eqn{E(Y|T = another treatment)}}

\item{\code{difft}}{vector of difference between twin1 and twin2}
}}
\section{Methods}{

\describe{
\item{\code{computeDifft()}}{Compute difference between twin1 and twin2. See details.}
}}
\seealso{
\code{\link{VT.forest}}, \code{\link{VT.forest.one}}, \code{\link{VT.forest.double}}
}

