% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\docType{class}
\name{VT.forest}
\alias{VT.forest}
\title{Difft by Random Forest}
\description{
An abstract reference class to compute twin via random forests
 
\code{VT.forest} extends \code{VT.difft}
}
\section{Fields}{

\describe{
\item{\code{...}}{see fields of \code{\link{VT.difft}}}
}}
\section{Methods}{

\describe{
\item{\code{checkModel(model)}}{Checking model class: Must be : train, RandomForest, randomForest}

\item{\code{getFullData()}}{Return twin1, twin2 and difft in column}

\item{\code{run()}}{Compute twin1 and twin2 estimation. Switch treatment if necessary.}
}}
\seealso{
\code{\link{VT.difft}}, \code{\link{VT.forest.one}}, \code{\link{VT.forest.double}}
}

