% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.R
\docType{class}
\name{VT.forest-class}
\alias{VT.forest}
\alias{VT.forest-class}
\title{A abstract reference class to compute twin via random forests}
\description{
A abstract reference class to compute twin via random forests
}
\section{Methods}{

\describe{
\item{\code{checkModel(model)}}{Checking model class: Must be : train, RandomForest, randomForest}

\item{\code{getFullData()}}{Return twin1, twin2 and difft in column}

\item{\code{run()}}{Compute twin1 and twin2 computation. Switch treatment if necessary.}
}}

