% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.double.R
\docType{class}
\name{VT.forest.double-class}
\alias{VT.forest.double}
\alias{VT.forest.double-class}
\title{A reference class to compute twins via double random forests}
\description{
A reference class to compute twins via double random forests
}
\section{Fields}{

\describe{
\item{\code{model_trt1}}{a caret/RandomForest/randomForest object for treatment T = 1}

\item{\code{model_trt0}}{a caret/RandomForest/randomForest object for treatment T = 0}
}}
\section{Methods}{

\describe{
\item{\code{computeTwin1()}}{Compute twin1 with OOB predictions from double forests}

\item{\code{computeTwin2()}}{Compute twin2 by the other part of data in the other forest}
}}

