% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.fold.R
\docType{class}
\name{VT.forest.fold-class}
\alias{VT.forest.fold}
\alias{VT.forest.fold-class}
\title{A reference class to compute twins via k random forest}
\description{
A reference class to compute twins via k random forest
}
\section{Fields}{

\describe{
\item{\code{interactions}}{logical set TRUE if model has been computed with interactions}

\item{\code{fold}}{numeric Number of fold, i.e. number of forest}

\item{\code{ratio}}{numeric}

\item{\code{groups}}{vector Define which observations belong to which group}
}}
\section{Methods}{

\describe{
\item{\code{run()}}{Compute twin1 and twin2 computation. Switch treatment if necessary.}
}}

