% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.one.R
\docType{class}
\name{VT.forest.one-class}
\alias{VT.forest.one}
\alias{VT.forest.one-class}
\title{A reference class to compute twins via one random forest}
\description{
A reference class to compute twins via one random forest
}
\section{Fields}{

\describe{
\item{\code{model}}{ANY a caret/RandomForest/randomForest class object}

\item{\code{interactions}}{logical set TRUE if model has been computed with interactions}

\item{\code{...}}{field from parent class : VT.forest}
}}
\section{Methods}{

\describe{
\item{\code{computeTwin1()}}{Compute twin1 with OOB predictions}

\item{\code{computeTwin2()}}{Compute twin2 by switching treatment and applying random forest model}
}}

