% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.wrapper.R
\name{vt.forest}
\alias{vt.forest}
\title{Create forest to compute difft}
\usage{
vt.forest(forest.type = "one", vt.data, interactions = T,
  method = "absolute", ...)
}
\arguments{
\item{forest.type}{character one / double / fold}

\item{vt.data}{\code{\link{VT.data}} or return of \code{vt.data()} function}

\item{interactions}{logical. If running VirtualTwins with treatment's
interactions, set to TRUE (default value)}

\item{method}{character absolute / relative / logit}

\item{...}{parameters of \code{\link{VT.difft}} or \code{\link{VT.forest}}}
}
\value{
\code{VT.difft}
}
\description{
\code{vt.forest} is a wrapper of \code{\link{VT.forest.one}},
\code{\link{VT.forest.double}} and \code{\link{VT.forest.fold}}.
}

