% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{VT.object-class}
\alias{VT.object}
\alias{VT.object-class}
\title{A Reference Class to deal with RCT dataset}
\description{
A Reference Class to deal with RCT dataset
}
\section{Fields}{

\describe{
\item{\code{data}}{A data.frame de la forme \eqn{Y,T,X_{1}, \ldots, X_{p}}. Y must
be two levels factor if type is binary. T must be numeric or integer.}

\item{\code{alpha}}{no usefull now, set to 1}

\item{\code{screening}}{logical, set to FALSE. Se TRUE to use varimp in trees
computation}

\item{\code{varimp}}{character vector of important variables to use in trees
computation}

\item{\code{delta}}{numeric representing the difference of incidence between
treatments}

\item{\code{type}}{character : binary or continous. Only binary is possible.}
}}
\section{Methods}{

\describe{
\item{\code{computeDelta()}}{Compute delta value.}

\item{\code{getFormula()}}{Return formula : Y~T+X1+...+Xp. Usefull for cforest function.}

\item{\code{getIncidences()}}{Return incidence table of data.}

\item{\code{getX(interactions = T, trt = NULL)}}{Return predictors {T,X,X*T,X*(1-T)}. Or {T,X} if interactions is FALSE.
If trt is not NULL, return predictors for T=trt}

\item{\code{getXwithInt()}}{Return predictors with interactions. Use VT.object::getX(interactions = T) instead.}

\item{\code{getY(trt = NULL)}}{Return outcome. If trt is not NULL, return outcome for T=trt.}

\item{\code{switchTreatment()}}{Switch treatment value.}
}}

