% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{vt.subgroups}
\alias{vt.subgroups}
\title{Visualize subgroups}
\usage{
vt.subgroups(vt.trees, only.leaf = T, only.fav = T, tables = F,
  verbose = F)
}
\arguments{
\item{vt.trees}{\code{\link{VT.tree}} object (can be a list)}

\item{only.leaf}{logical select only leaf of trees}

\item{only.fav}{logical select only favorable subgroup (meaning with favorable label of the tree)}

\item{tables}{logical show tables of incidence}

\item{verbose}{print tables during computation}
}
\value{
data.frame of rules
}
\description{
Visualize subgroups
}

