% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\docType{class}
\name{VT.tree-class}
\alias{VT.tree}
\alias{VT.tree-class}
\title{An abstract reference class to compute tree}
\description{
An abstract reference class to compute tree
}
\section{Fields}{

\describe{
\item{\code{vt.difft}}{VT.difft object}

\item{\code{outcome}}{vector}

\item{\code{threshold}}{numeric Threshold for difft (c)}

\item{\code{screening}}{logical TRUE if using varimp (default is VT.object screening field)}

\item{\code{sens}}{character Sens can be ">" (default) or "<". Meaning : difft > threshold or difft < threshold}

\item{\code{name}}{character Names of the tree}

\item{\code{tree}}{rpart Rpart object to construct the tree}

\item{\code{Ahat}}{vector Indicator of beglonging to Ahat}
}}

