% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.class.R
\docType{class}
\name{VT.tree.class-class}
\alias{VT.tree.class}
\alias{VT.tree.class-class}
\title{A reference class to compute subgroups by classifiation tree}
\description{
A reference class to compute subgroups by classifiation tree
}

