% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.class.R
\docType{class}
\name{VT.tree.class}
\alias{VT.tree.class}
\title{Classification tree to find subgroups}
\description{
See \code{\link{VT.tree}}
}
\section{Methods}{

\describe{
\item{\code{run(...)}}{Compute tree with rpart parameters}
}}

