% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.wrapper.R
\name{vt.tree}
\alias{vt.tree}
\title{Trees to find Subgroups}
\usage{
vt.tree(tree.type = "class", vt.difft, sens = ">", threshold = seq(0.5,
  0.8, 0.1), screening = NULL, ...)
}
\arguments{
\item{tree.type}{character "class" for classification tree, "reg" for regression tree}

\item{vt.difft}{\code{\link{VT.difft}} object}

\item{sens}{character c(">","<"). See details.}

\item{threshold}{numeric It can be a unique value or a vector}
}
\value{
\code{VT.tree} or a list of \code{VT.tree} depending on threshold dimension
}
\description{
A wrapper of class VT.tree.xxx
}
\details{
See \code{\link{VT.tree}}
}

