% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.reg.R
\docType{class}
\name{VT.tree.reg}
\alias{VT.tree.reg}
\title{Regression tree to find subgroups}
\description{
See \code{\link{VT.tree}}
}
\section{Methods}{

\describe{
\item{\code{run(...)}}{Compute tree with rpart parameters}
}}

