\name{VirtualTwins-package}
\alias{VirtualTwins-package}
\alias{VirtualTwins}
\docType{package}
\title{
\packageTitle{VirtualTwins}
}
\description{
\packageDescription{VirtualTwins}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VirtualTwins}
\packageIndices{VirtualTwins}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VirtualTwins}

Maintainer: \packageMaintainer{VirtualTwins}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
